requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.30.0
transformers==4.45.0
datasets==2.19.0
torch==2.2.1
pandas==2.2.2
PyMuPDF==1.23.26
python-docx==1.1.0
openpyxl==3.1.2
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
PyJWT==2.9.0
jinja2==3.1.4
requests==2.32.3
python-multipart==0.0.9
python-dotenv==1.0.1
pytest==8.2.1
httpx==0.27.0
pydantic-settings==2.4.0
sentence-transformers==2.7.0
faiss-cpu==1.8.0
